requirements/base.txt
asgiref==3.4.*
Django==3.2.*
pytz==2021.*
sqlparse==0.4.*
Pillow==9.1.*
django-ranged-response==0.2.*
django-simple-captcha==0.5.*
psycopg2-binary==2.9.*
redis==4.1.*
celery==5.2.*
requests==2.27.*
django-tinymce==3.4.*
django-jalali==6.0.*
django-mptt==0.13.*
az-iranian-bank-gateways==1.6.*
